refactor(front): move triggers DTOs into types/api and rename Codec->Schema
configuration/triggers.ts was already zod (the *Codec names were misnomers on
z.object schemas), so there was no io-ts to migrate. Renamed the three
*RequestBodyCodec schemas to *RequestBodySchema (GEN13) and moved all transport
DTOs/schemas to types/api/assistant/configuration/triggers.ts.

- Moved: GetUserTriggersResponseBody, GetTriggersResponseBody(+Schema),
  Delete/Patch/PostTriggersRequestBody(+Schema), PostTextAsCronRuleResponseBody,
  PostTextAsCronRuleRequestBody(+Schema).
- Kept in lib: generateScheduleRule, validateIntervalConfig, IANA/cron helpers,
  CRON_REGEXP, and the error-message constants.
- Codec->Schema rename had zero external consumers (pure internal rename).
- 6 consumers updated; text_as_cron_rule route split (DTO -> types, constants +
  generateScheduleRule stay on lib).
