+++ front-api/routes/w/[wId]/me/triggers.ts
@@ -1,6 +1,6 @@
 import { getAgentConfigurations } from "@app/lib/api/assistant/configuration/agent";
-import type { GetUserTriggersResponseBody } from "@app/lib/api/assistant/configuration/triggers";
 import { TriggerResource } from "@app/lib/resources/trigger_resource";
+import type { GetUserTriggersResponseBody } from "@app/types/api/assistant/configuration/triggers";
 import { removeNulls } from "@app/types/shared/utils/general";
 import { workspaceApp } from "@front-api/middlewares/ctx";
 import type { HandlerResult } from "@front-api/middlewares/utils";
