+++ front/components/agent_builder/submitAgentBuilderForm.ts
@@ -8,10 +8,6 @@ import {
   getTableIdForContentNode,
 } from "@app/components/agent_builder/shared/tables";
 import type { AdditionalConfigurationInBuilderType } from "@app/components/shared/tools_picker/types";
-import type {
-  PatchTriggersRequestBody,
-  PostTriggersRequestBody,
-} from "@app/lib/api/assistant/configuration/triggers";
 import type { TableDataSourceConfiguration } from "@app/lib/api/assistant/configuration/types";
 import type {
   GetContentNodesOrChildrenRequestBodyType,
@@ -31,6 +27,10 @@ import type {
   DataSourcesConfigurationsCodecType,
   PostOrPatchAgentConfigurationRequestBody,
 } from "@app/types/api/agent_configuration";
+import type {
+  PatchTriggersRequestBody,
+  PostTriggersRequestBody,
+} from "@app/types/api/assistant/configuration/triggers";
 import type {
   AgentConfigurationType,
   LightAgentConfigurationType,
