+++ front/lib/api/assistant/configuration/triggers.ts
@@ -1,80 +1,8 @@
 import { getCronTimezoneGeneration } from "@app/lib/api/assistant/configuration/triggers/cron_timezone";
 import type { Authenticator } from "@app/lib/auth";
-import type {
-  ScheduleConfig,
-  TriggerType,
-} from "@app/types/assistant/triggers";
-import {
-  FullTriggerSchema,
-  TriggerSchema,
-} from "@app/types/assistant/triggers";
+import type { ScheduleConfig } from "@app/types/assistant/triggers";
 import type { Result } from "@app/types/shared/result";
 import { Err, Ok } from "@app/types/shared/result";
-import { z } from "zod";
-
-export interface GetUserTriggersResponseBody {
-  triggers: (TriggerType & {
-    isEditor: boolean;
-    agentName: string;
-    agentPictureUrl: string;
-  })[];
-}
-
-export const GetTriggersResponseBodySchema = z.object({
-  triggers: z.array(
-    FullTriggerSchema.and(
-      z.object({
-        isEditor: z.boolean(),
-        editorName: z.string().optional(),
-      })
-    )
-  ),
-});
-export type GetTriggersResponseBody = z.infer<
-  typeof GetTriggersResponseBodySchema
->;
-
-export const DeleteTriggersRequestBodyCodec = z.object({
-  triggerIds: z.array(z.string()),
-});
-export type DeleteTriggersRequestBody = z.infer<
-  typeof DeleteTriggersRequestBodyCodec
->;
-
-export const PatchTriggersRequestBodyCodec = z.object({
-  triggers: z.array(z.object({ sId: z.string() }).and(TriggerSchema)),
-});
-export type PatchTriggersRequestBody = z.infer<
-  typeof PatchTriggersRequestBodyCodec
->;
-
-export const PostTriggersRequestBodyCodec = z.object({
-  triggers: z.array(TriggerSchema),
-});
-export type PostTriggersRequestBody = z.infer<
-  typeof PostTriggersRequestBodyCodec
->;
-
-// Backward-compatible: cron responses include cronRule, interval responses
-// include interval fields.
-export type PostTextAsCronRuleResponseBody =
-  | { type?: "cron"; cronRule: string; timezone: string }
-  | {
-      type: "interval";
-      intervalDays: number;
-      dayOfWeek: number | null;
-      hour: number;
-      minute: number;
-      timezone: string;
-    };
-
-export const PostTextAsCronRuleRequestBodySchema = z.object({
-  naturalDescription: z.string(),
-  defaultTimezone: z.string(),
-});
-export type PostTextAsCronRuleRequestBody = z.infer<
-  typeof PostTextAsCronRuleRequestBodySchema
->;
 
 function isValidIANATimezone(timezone: string): boolean {
   // Get the list of all supported IANA timezones
